Рефакторинг модуля со вспомогательными функциями
