+++ src/utils/utils.js
@@ -34,14 +34,4 @@ const calculateDuration = (dateFrom, dateTo) => {
   return { hours, minutes };
 };
 
-
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (elements) => elements[Math.floor(Math.random() * elements.length)];
-
-export { getRandomInteger, getRandomArrayElement, formatDate, calculateDuration };
+export { formatDate, calculateDuration };
