+++ src/view/edit-point-view.js
@@ -1,5 +1,5 @@
 import { EVENT_TYPES } from '../const.js';
-import { formatDate } from '../utils.js';
+import { formatDate } from '../utils/utils.js';
 import AbstractView from '../framework/view/abstract-view.js';
 
 function createOffersTemplate(offersList, offers) {
